Made the vector forward and back projections double precision Same thing for the conjugate gradient operator, and the applications involved

+++ applications/rtkvectorbackprojections/rtkvectorbackprojections.cxx
@@ -40,7 +40,7 @@ int main(int argc, char * argv[])
 {
   GGO(rtkvectorbackprojections, args_info);
 
-  typedef float OutputPixelType;
+  typedef double OutputPixelType;
   const unsigned int Dimension = 3;
 #ifdef RTK_USE_CUDA
   typedef itk::CudaImage< OutputPixelType, Dimension > OutputImageType;
