+++ code/rtkIterativeConeBeamReconstructionFilter.cxx
@@ -23,20 +23,20 @@ namespace rtk
 
 
 template<>
-typename rtk::BackProjectionImageFilter<itk::VectorImage<float, 3>, itk::VectorImage<float, 3> >::Pointer
-IterativeConeBeamReconstructionFilter< itk::VectorImage<float, 3>, itk::VectorImage<float, 3> >
+typename rtk::BackProjectionImageFilter<itk::VectorImage<double, 3>, itk::VectorImage<double, 3> >::Pointer
+IterativeConeBeamReconstructionFilter< itk::VectorImage<double, 3>, itk::VectorImage<double, 3> >
 ::InstantiateBackProjectionFilter(int bptype)
 {
-  rtk::BackProjectionImageFilter< itk::VectorImage<float, 3>, itk::VectorImage<float, 3> >::Pointer bp;
+  rtk::BackProjectionImageFilter< itk::VectorImage<double, 3>, itk::VectorImage<double, 3> >::Pointer bp;
   switch(bptype)
     {
 //    case(0):
 //      bp = rtk::BackProjectionImageFilter<ProjectionStackType, VolumeType>::New();
 //      break;
     case(1):
-      bp = rtk::JosephBackProjectionImageFilter<itk::VectorImage<float, 3>,
-              itk::VectorImage<float, 3>,
-              Functor::SplatWeightMultiplication< float, double, float > >::New();
+      bp = rtk::JosephBackProjectionImageFilter<itk::VectorImage<double, 3>,
+              itk::VectorImage<double, 3>,
+              Functor::SplatWeightMultiplication< double, double, double > >::New();
       break;
 //    case(2):
 //    #ifdef RTK_USE_CUDA
@@ -62,18 +62,18 @@ IterativeConeBeamReconstructionFilter< itk::VectorImage<float, 3>, itk::VectorIm
 }
 
 template<>
-typename rtk::ForwardProjectionImageFilter<itk::VectorImage<float, 3>, itk::VectorImage<float, 3> >::Pointer
-IterativeConeBeamReconstructionFilter< itk::VectorImage<float, 3>, itk::VectorImage<float, 3> >
+typename rtk::ForwardProjectionImageFilter<itk::VectorImage<double, 3>, itk::VectorImage<double, 3> >::Pointer
+IterativeConeBeamReconstructionFilter< itk::VectorImage<double, 3>, itk::VectorImage<double, 3> >
 ::InstantiateForwardProjectionFilter(int fwtype)
 {
-  rtk::ForwardProjectionImageFilter<itk::VectorImage<float, 3>, itk::VectorImage<float, 3> >::Pointer fw;
+  rtk::ForwardProjectionImageFilter<itk::VectorImage<double, 3>, itk::VectorImage<double, 3> >::Pointer fw;
   switch(fwtype)
     {
     case(0):
-      fw = rtk::JosephForwardProjectionImageFilter<itk::VectorImage<float, 3>,
-              itk::VectorImage<float, 3>,
-              Functor::VectorInterpolationWeightMultiplication<float, double, itk::VariableLengthVector<float>>,
-              Functor::VectorProjectedValueAccumulation<itk::VariableLengthVector<float>, itk::VariableLengthVector<float> > >::New();
+      fw = rtk::JosephForwardProjectionImageFilter<itk::VectorImage<double, 3>,
+              itk::VectorImage<double, 3>,
+              Functor::VectorInterpolationWeightMultiplication<double, double, itk::VariableLengthVector<double>>,
+              Functor::VectorProjectedValueAccumulation<itk::VariableLengthVector<double>, itk::VariableLengthVector<double> > >::New();
     break;
 //    case(1):
 //      fw = rtk::RayCastInterpolatorForwardProjectionImageFilter<VolumeType, ProjectionStackType>::New();
